# Fix Welcome Page Performance Issues-created-by-agentic