--- a/src/main/resources/application.properties
+++ b/src/main/resources/application.properties
@@ -6,8 +6,13 @@ database=h2
 spring.sql.init.schema-locations=classpath*:db/${database}/schema.sql
 spring.sql.init.data-locations=classpath*:db/${database}/data.sql
 
-# Web
+# Template engine optimization
+spring.thymeleaf.cache=true
 spring.thymeleaf.mode=HTML
+spring.thymeleaf.template-resolver-order=1
+spring.thymeleaf.enable-spring-el-compiler=true
+spring.thymeleaf.prefix=classpath:/templates/
+spring.thymeleaf.suffix=.html
 
 # JPA
 spring.jpa.hibernate.ddl-auto=none
@@ -27,5 +32,12 @@ logging.level.org.springframework=INFO
 # logging.level.org.springframework.web=DEBUG
 # logging.level.org.springframework.context.annotation=TRACE
 
-# Maximum time static resources should be cached
+# Static resources caching
 spring.web.resources.cache.cachecontrol.max-age=12h
+spring.resources.chain.strategy.content.enabled=true
+spring.resources.chain.strategy.content.paths=/**
+
+# Server compression
+server.compression.enabled=true
+server.compression.mime-types=text/html,text/css,application/javascript,application/json
+server.compression.min-response-size=2048