This PR addresses the FreezeException and performance issues in the PetClinic welcome page endpoint (/). The changes include:

1. Template Engine Optimizations:
   - Enabled Thymeleaf caching
   - Configured template resolver order
   - Enabled Spring EL compiler
   - Added static resource caching
   - Enabled server compression

2. Controller Improvements:
   - Added response caching using @Cacheable
   - Optimized ModelAndView usage
   - Reduced memory allocations

Performance Impact:
- Current P50 latency: 396.21ms → Expected: ~100ms
- Current P95 latency: 413.19ms → Expected: ~150ms
- Rendering time: 273.16ms (67.56%) → Expected: ~50ms

Testing:
- Verified template caching works
- Confirmed response compression
- Tested memory usage under load

The changes should significantly reduce the rendering time and prevent FreezeException occurrences.